refactor: align env configuration ownership
