+++ graphql/server/src/middleware/api.ts
@@ -1,4 +1,4 @@
-import { getNodeEnv } from '@pgpmjs/env';
+import { getNodeEnv } from '@constructive-io/graphql-env';
 import { Logger } from '@pgpmjs/logger';
 import { svcCache } from '@pgpmjs/server-utils';
 import { parseUrl } from '@constructive-io/url-domains';
