+++ graphql/server/src/middleware/error-handler.ts
@@ -1,4 +1,4 @@
-import { getNodeEnv } from '@pgpmjs/env';
+import { getNodeEnv } from '@constructive-io/graphql-env';
 import { Logger } from '@pgpmjs/logger';
 import type { ErrorRequestHandler, NextFunction, Request, Response } from 'express';
