+++ pgpm/types/src/index.ts
@@ -3,5 +3,4 @@ export * from './error';
 export * from './error-factory';
 export * from './pg-error-format';
 export * from './pgpm';
-export * from './jobs';
 export * from './update';
