+++ uploads/s3-streamer/__tests__/uploads.test.ts
@@ -1,5 +1,5 @@
 import { S3Client } from '@aws-sdk/client-s3';
-import { getEnvOptions } from '@pgpmjs/env';
+import { getEnvOptions } from '@constructive-io/graphql-env';
 import { createS3Bucket } from '@constructive-io/s3-utils';
 import { createReadStream } from 'fs';
 import { sync as glob } from 'glob';
