+++ uploads/s3-streamer/package.json
@@ -29,7 +29,7 @@
     "test:watch": "jest --watch"
   },
   "devDependencies": {
-    "@pgpmjs/env": "workspace:^",
+    "@constructive-io/graphql-env": "workspace:^",
     "glob": "^13.0.6",
     "makage": "^0.3.0"
   },
@@ -37,8 +37,8 @@
     "@aws-sdk/client-s3": "^3.1052.0",
     "@aws-sdk/lib-storage": "^3.1052.0",
     "@constructive-io/content-type-stream": "workspace:^",
-    "@constructive-io/s3-utils": "workspace:^",
-    "@pgpmjs/types": "workspace:^"
+    "@constructive-io/graphql-types": "workspace:^",
+    "@constructive-io/s3-utils": "workspace:^"
   },
   "keywords": [
     "s3",
